move VRT_call to cache_vrt_vcl.c for next commit
