+++ bin/varnishd/cache/cache_vrt_vcl.c
@@ -42,6 +42,7 @@
 
 #include "cache_director.h"
 #include "cache_vcl.h"
+#include "vcc_interface.h"
 
 /*--------------------------------------------------------------------*/
 
@@ -477,3 +478,47 @@ VCL_##func##_method(struct vcl *vcl, struct worker *wrk,		\
 }
 
 #include "tbl/vcl_returns.h"
+
+/*--------------------------------------------------------------------
+ */
+
+static void
+no_rollback(void *priv)
+{
+	VRT_CTX;
+
+	CAST_OBJ_NOTNULL(ctx, priv, VRT_CTX_MAGIC);
+	VRT_fail(ctx, "no rollback during dynamic sub calls");
+}
+
+VCL_VOID
+VRT_call(VRT_CTX, VCL_SUB sub)
+{
+	struct vmod_priv *p;
+
+	CHECK_OBJ_NOTNULL(ctx, VRT_CTX_MAGIC);
+
+	p = VRT_priv_task(ctx, (void *)sub->func);
+	if (p == NULL) {
+		VRT_fail(ctx, "no priv task - out of ws?");
+		return;
+	}
+
+	if (p->priv != NULL) {
+		assert(p->priv == ctx);
+		assert(p->free == no_rollback);
+		VRT_fail(ctx, "recursive call to %s", sub->name);
+		return;
+	}
+
+	p->priv = TRUST_ME(ctx);
+	p->free = no_rollback;
+
+	if (sub->methods & ctx->method)
+		sub->func(ctx);
+	else
+		VRT_fail(ctx, "not allowed here");
+
+	p->priv = NULL;
+	p->free = NULL;
+}
